refactor: create `user-menu.tsx`
